+++ HISTORY.md
@@ -1,5 +1,9 @@
 # Revision History
 
+## Revision 0.5.1 (WSM)
+- Fixed lots of compiler errors in sp_hnfs.f90 and grid_utils.f90.
+- Added grid_utils.f90 to the Makefile.
+
 ## Revision 0.5.0 (WSM)
 - Changed sp_hnfs.f90 to use integers instead of reals.
 - Changed sp_hnfs.f90 to return only one grid per determinant size